reading the bugdet list in console

+++ index.js
@@ -1,5 +1,7 @@
 import fs from "fs";
 import { Command } from "commander";
+import chalk from "chalk";
+
 const program = new Command();
 
 program
@@ -33,12 +35,40 @@ program
       budget = [];
     }
     budget = JSON.parse(loadedBudget);
+    const budgetExists = budget.find(
+      (currentBudget) => currentBudget.title === title,
+    );
+    if (budgetExists) {
+      console.log(chalk.bgBlue(`budget with similar '${title}' exists`));
+      return;
+    }
 
     budget.push(newBudget);
     //const budgetArrays = [];
     //budgetArrays.push(newBudget);
     //console.log(budgetArrays)
     fs.writeFileSync("./data/budget.json", JSON.stringify(budget));
+    console.log(chalk.bgMagentaBright(`bugdet added succefully`));
   });
+
+  //reads all bugdet to the console
+  program.command("read")
+  .description("lists all the budget listed")
+  .action(function () {
+    const loadedBudget = fs.readFileSync("./data/budget.json", "utf-8");
+    const budget = JSON.parse(loadedBudget);
+
+    if(budget.length === 0) {
+        console.log(chalk.bgCyanBright(`no any bugdet added,, add one soon!!`))
+        return;
+    }
+
+    budget.forEach(function(currentBudget) {
+        console.log(currentBudget.title)
+        console.log(currentBudget.quantity)
+        console.log(currentBudget.price)
+        console.log("_____________________")
+    })
+  })
 
 program.parse(process.argv);
